# vouch-world/openapi/read.yaml
openapi: 3.1.0
info:
  title: vouch — read-only observation API
  version: "1.0.0"
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  summary: The read-only HTTP connection point of the vouch world (Layer 5 Observation).
  description: |
    External clients **connect here to watch** the deterministic vouch world: its
    state, regions, agents, the event log, and derived metrics. This surface is
    **GET-only by construction** — the server is handed only a read-only `WorldView`
    (no `emit` / `run` / `advanceTick`), so "watching cannot write" is enforced by the
    type system, not by discipline (§2-6, §2-9).

    Generated from the real implementation: `vouch-world/src/observation/server.ts` +
    `metrics.ts`, with response schemas derived from the real TypeScript types
    (`RegionState`, `AgentState`, `AlmaEvent`, `Metrics`, `Certificate`).

    Notes for clients:
      - **No write surface exists here.** Writing goes through `vouch-node` — the
        authenticated participate node (`POST /v1/register`, `POST /v1/command`), which
        mounts this same read-only observation app for its GETs.
      - **No CORS headers are set** by the server today. A browser app on another origin
        must use a dev proxy (or have CORS configured at deploy time).
      - The reference example (`examples/observe.ts`) runs a finite 8-tick simulation and
        then serves a **static** world, so `tick` does not advance while the server is up.
servers:
...
paths:
...
  /regions/{id}:
    get:
      tags: [state]
      operationId: getRegion
      summary: A single region by id.
      parameters:
        - name: id
          in: path
          required: true
          description: The region id (the `region` string, e.g. "umi").
          schema: { type: string }
      responses:
        "200":
          description: The region.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegionState" }
        "404":
          description: No region with that id.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "region not found" }
...
  /agents/{id}:
    get:
      tags: [state]
      operationId: getAgent
      summary: A single agent by id.
      parameters:
        - name: id
          in: path
          required: true
          description: The agent id (`name@region`, e.g. "ada@umi").
          schema: { type: string }
      responses:
        "200":
          description: The agent.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentState" }
        "404":
          description: No agent with that id.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "agent not found" }
...
  /items/{id}:
    get:
      tags: [state]
      operationId: getItem
      summary: A single item by id.
      parameters:
        - name: id
          in: path
          required: true
          description: The item id chosen at mint, e.g. "deed-umi-1".
          schema: { type: string }
      responses:
        "200":
          description: The item.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ItemState" }
        "404":
          description: No item with that id.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "item not found" }
...
  /definitions/{id}:
    get:
      tags: [state]
      operationId: getDefinition
      summary: A single definition by id.
      parameters:
        - name: id
          in: path
          required: true
          description: The definition id (`<namespace>.<name>`, e.g. "core.transfer").
          schema: { type: string }
      responses:
        "200":
          description: The definition.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DefinitionRecord" }
        "404":
          description: No definition with that id.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "definition not found" }
  /log:
    get:
      tags: [log]
      operationId: getLog
      summary: Events from the append-only log, optionally since a seq.
      description: >-
        Returns at most 1000 events per call. A caller that receives exactly 1000 should
        ask again from `since + 1000`; `/log/digest` reports the total length.
      parameters:
        - name: since
          in: query
          required: false
          description: >-
            Return events with `seq >= since`. Defaults to 0. MUST be a non-negative
            integer — anything else is a 400 rather than being coerced, because a
            coerced value silently returned the whole log.
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: The events at or after `since`, in seq order, capped at 1000.
          content:
            application/json:
              schema:
                type: array
                maxItems: 1000
                items: { $ref: "#/components/schemas/AlmaEvent" }
        "400":
          description: '`since` was not a non-negative integer.'
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "since must be a non-negative integer" }
...
components:
  schemas:
    Tick:
      type: integer
      minimum: 0
      description: Discrete sim-engine time. The protocol orders by log `seq`, not `tick` (audit G5).
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }

    # --- derived metrics (Layer 5) ------------------------------------------
...
    AgentRole:
      type: string
      enum: [artisan, merchant, broker, treasury]
    ValueProfile:
      type: string
      enum: [strict, lenient]
    RecognitionStatus:
      type: string
      enum: [unrecognized, recognized]
    ForeignCertStance:
      type: string
      description: How a village translates another village's certificates (§4-A).
      enum: [absorb, map, reexamine, reject]

    # --- region (Layer 2) ---------------------------------------------------
    SchemaLedgerEntry:
      type: object
      required: [schemaId]
      properties:
        schemaId: { type: string }
        label: { type: string }
    VerificationPolicy:
      type: object
      required: [acceptedSchemaIds, rejectUnknownSchemas]
      properties:
        acceptedSchemaIds:
          type: array
          items: { type: string }
        rejectUnknownSchemas: { type: boolean }
    DiplomacyPolicy:
      type: object
      required: [defaultStance, overrides]
      properties:
        defaultStance: { $ref: "#/components/schemas/ForeignCertStance" }
        overrides:
          type: object
          additionalProperties: { $ref: "#/components/schemas/ForeignCertStance" }
    Governance:
      description: Who may amend a region — its constitution (§8). dictatorship → the owner; council → listed members + a vote threshold.
      oneOf:
        - type: object
          required: [kind]
          properties: { kind: { type: string, const: dictatorship } }
        - type: object
          required: [kind, members, threshold]
          properties:
            kind: { type: string, const: council }
            members: { type: array, items: { type: string } }
            threshold: { type: integer, minimum: 1 }
    EconomyPolicy:
      type: object
      description: The region's own fee/credit schedule (read by executeTransfer for the sender's region).
      required: [baseCostRate, minCostRate, repDiscount, creditPerTx]
      properties:
        baseCostRate: { type: number, minimum: 0, maximum: 1 }
        minCostRate: { type: number, minimum: 0, maximum: 1 }
        repDiscount: { type: number, minimum: 0 }
        creditPerTx: { type: integer, minimum: 0 }
    ResourcePolicy:
      type: object
      description: A region's finite resource pool config (P3 scarcity, the "compete" substrate).
      required: [capacity, regenPerTick]
      properties:
        capacity: { type: integer, minimum: 0 }
        regenPerTick: { type: integer, minimum: 0 }
    RegionLifecycle:
      type: string
      description: active = running; dormant = hibernated by the owner (can be listed + sold). A region is never deleted.
      enum: [active, dormant]
    InstitutionChange:
      description: A logged amendment to one institution policy, tagged by `policy`.
      oneOf:
        - { type: object, required: [policy, value], properties: { policy: { type: string, const: verification }, value: { $ref: "#/components/schemas/VerificationPolicy" } } }
        - { type: object, required: [policy, value], properties: { policy: { type: string, const: diplomacy }, value: { $ref: "#/components/schemas/DiplomacyPolicy" } } }
        - { type: object, required: [policy, value], properties: { policy: { type: string, const: schemaLedger }, value: { type: array, items: { $ref: "#/components/schemas/SchemaLedgerEntry" } } } }
        - { type: object, required: [policy, value], properties: { policy: { type: string, const: governance }, value: { $ref: "#/components/schemas/Governance" } } }
        - { type: object, required: [policy, value], properties: { policy: { type: string, const: economy }, value: { $ref: "#/components/schemas/EconomyPolicy" } } }
        - { type: object, required: [policy, value], properties: { policy: { type: string, const: resource }, value: { $ref: "#/components/schemas/ResourcePolicy" } } }
        - { type: object, required: [policy, value], properties: { policy: { type: string, const: items }, value: { $ref: "#/components/schemas/ItemPolicy" } } }
    GovProposal:
      type: object
      description: A council's one in-flight amendment vote (P3); applies once `votes` reach the threshold.
      required: [change, votes, proposedBy]
      properties:
        change: { $ref: "#/components/schemas/InstitutionChange" }
        votes: { type: array, items: { type: string } }
        proposedBy: { type: string }
    ItemPolicy:
      type: object
      description: >-
        Who may mint items for this region's residents. Optional in live state:
        regions replayed from journals older than the policy never carried it, and
        readers fall back to {minting: owner}.
      required: [minting]
      properties:
        minting:
          type: string
          enum: [owner, residents, anyone]
    Institutions:
      type: object
      required: [schemaLedger, verificationPolicy, diplomacyPolicy, governance, economyPolicy, resourcePolicy]
      properties:
        schemaLedger:
          type: array
          items: { $ref: "#/components/schemas/SchemaLedgerEntry" }
        verificationPolicy: { $ref: "#/components/schemas/VerificationPolicy" }
        diplomacyPolicy: { $ref: "#/components/schemas/DiplomacyPolicy" }
        governance: { $ref: "#/components/schemas/Governance" }
        economyPolicy: { $ref: "#/components/schemas/EconomyPolicy" }
        resourcePolicy: { $ref: "#/components/schemas/ResourcePolicy" }
        itemPolicy: { $ref: "#/components/schemas/ItemPolicy" }
    Proposer:
      description: Who proposed a region's founding.
      oneOf:
        - type: object
          required: [kind]
          properties:
            kind: { type: string, const: genesis }
        - type: object
          required: [kind]
          properties:
            kind: { type: string, const: experimenter }
            note: { type: string }
        - type: object
          required: [kind, sourceRegion, reason, cohort]
          properties:
            kind: { type: string, const: emergence }
            sourceRegion: { type: string }
            reason: { type: string }
            cohort:
              type: array
              items: { type: string }
    RegionState:
      type: object
      required: [id, displayName, institutions, status, proposer, foundedAtSeq, owner, lifecycle, salePrice, openProposal, resourceLevel]
      properties:
        id: { type: string }
        displayName: { type: string }
        institutions: { $ref: "#/components/schemas/Institutions" }
        status: { $ref: "#/components/schemas/RecognitionStatus" }
        proposer: { $ref: "#/components/schemas/Proposer" }
        foundedAtSeq:
          type: integer
          minimum: 0
          description: The log seq at which the region was founded (NOT the sim tick).
        owner:
          type: [string, "null"]
          description: The account/ID that governs this region; null = system/unowned (genesis, emergence).
        lifecycle: { $ref: "#/components/schemas/RegionLifecycle" }
        salePrice:
          type: [integer, "null"]
          description: Asking price when listed on the market; null = not for sale.
        openProposal:
          description: The council's one in-flight amendment vote (P3); null when none.
          oneOf:
            - { $ref: "#/components/schemas/GovProposal" }
            - { type: "null" }
        resourceLevel:
          type: integer
          minimum: 0
          description: Current amount in the region's resource pool (P3 scarcity); born 0.
      example:
        id: umi
        displayName: Umi
        institutions:
          schemaLedger: []
          verificationPolicy: { acceptedSchemaIds: [], rejectUnknownSchemas: false }
          diplomacyPolicy: { defaultStance: reexamine, overrides: {} }
          governance: { kind: dictatorship }
          economyPolicy: { baseCostRate: 0.2, minCostRate: 0.05, repDiscount: 0.02, creditPerTx: 1 }
          resourcePolicy: { capacity: 0, regenPerTick: 0 }
        status: recognized
        proposer: { kind: genesis }
        foundedAtSeq: 0
        owner: null
        lifecycle: active
        salePrice: null
        openProposal: null
        resourceLevel: 0

    # --- agent (Layer 3) ----------------------------------------------------
    Balances:
      type: object
      required: [credit, currency]
      properties:
        credit:
          type: number
          description: Non-transferable, slow trust accrual (§3-B).
        currency:
          type: number
          description: Transferable medium of exchange (§3-B).
    AgentState:
      type: object
      required: [id, region, role, publicKey, balances, reputation, trust, resources, valueProfile]
      properties:
        id:
          type: string
          description: name@region — stable identity across migration.
        region:
          type: string
          description: Current residence (changes on migration).
        role: { $ref: "#/components/schemas/AgentRole" }
        publicKey:
          type: string
          description: base64 Ed25519 public key ("" for a treasury account).
        balances: { $ref: "#/components/schemas/Balances" }
        reputation:
          type: number
          description: Economy-derived standing (accrues on settled trades).
        trust:
          type: number
          description: Social capital from being vouched for — distinct from reputation (a vouch doesn't buy a cheaper fee).
        resources:
          type: number
          description: Amount drawn from region resource pools (P3 scarcity).
        valueProfile: { $ref: "#/components/schemas/ValueProfile" }
    ItemState:
      type: object
      description: A unique, tradeable digital item (the ownership ledger), distinct from currency.
      required: [id, kind, owner]
      properties:
        id: { type: string }
        kind: { type: string }
        owner: { type: string, description: the agent (name@region) that holds it }

    # --- composed world (Layer 4) -------------------------------------------
...
    DefinitionRecord:
      type: object
      description: One versioned, log-stored definition (RFC 0007 §4). The engine validates only this envelope; `body` is never interpreted here.
      required: [kind, id, version, status, body]
      properties:
        kind:
          type: string
          description: Definition kind (e.g. `command`, `role`, `procedure`, `law`, `preamble`) — opaque to the engine.
          example: command
        id:
          type: string
          description: Namespaced id, `<namespace>.<name>`. `kernel.*` is reserved; `core.*` is seed-once.
          example: core.transfer
        version:
          type: integer
          minimum: 1
          description: Monotonic per id — a redefinition is always a new version (no shadowing, no gaps).
        status:
          type: string
          enum: [active, retired]
        body:
          type: object
          description: Opaque definition body; interpreted by vouch-node (preconditions/effects), never by the engine.
          additionalProperties: true

    # --- event log (Foundation B) -------------------------------------------
    AlmaEvent:
      type: object
      required: [seq, tick, type, actor, payload]
      properties:
        seq:
          type: integer
          minimum: 0
          description: Global, 0-based, monotonic position in the log (the canonical order).
        tick: { $ref: "#/components/schemas/Tick" }
        type:
          type: string
          description: |
            Event kind. Known kinds: region.founded, region.institution.changed,
            region.recognized, region.lifecycle.changed, region.listed,
            region.ownership.transferred, gov.proposal.opened, gov.vote.cast,
            resource.regenerated, resource.drawn, agent.admitted, agent.migrated,
            agent.decided, agent.vouched, economy.settled, economy.minted,
            item.minted, item.transferred, system.tick.
        actor:
          type: string
          description: The originating subject (an identifier), or "world" for the engine.
        payload:
          description: |
            Event content, opaque to the log itself; meaning belongs to higher layers.
            The richest payload is `economy.settled`, which carries a `receipt` Certificate
            and the signed balance deltas (see SettlementPayload). Other event kinds carry
            their own shapes, so a generic object is the fallback branch.
          oneOf:
            - $ref: "#/components/schemas/SettlementPayload"
            - type: object
              additionalProperties: true
      example:
        seq: 2
        tick: 0
        type: region.founded
        actor: world
        payload:
          status: recognized
          proposer: { kind: genesis }

    # --- settlement payload (the economy.settled event content) -------------
    SettlementEntry:
      type: object
      required: [agentId, currencyDelta, creditDelta, reputationDelta]
      properties:
        agentId: { type: string }
        currencyDelta: { type: number }
        creditDelta: { type: number }
        reputationDelta: { type: number }
    SettlementPayload:
      type: object
      description: The payload of an `economy.settled` event, authored by the environment.
      required: [entries, receipt, memo]
      properties:
        entries:
          type: array
          items: { $ref: "#/components/schemas/SettlementEntry" }
        receipt: { $ref: "#/components/schemas/Certificate" }
        memo:
          type: object
          required: [from, to, amount, fee]
          properties:
            from: { type: string }
            to: { type: string }
            amount: { type: integer }
            fee: { type: integer }

    # --- certificate (Layer 1 vouch-core envelope; appears inside payloads) --
    Certificate:
      type: object
      description: |
        The universal vouch-core certificate envelope. Not returned as a top-level body by
        any read route, but embedded in `economy.settled` event payloads as the transaction
        receipt (schemaId "alma.tx/receipt/v1").
      required: [version, suite, issuer, subject, schemaId, claims, issuedAt, signature]
      properties:
        version: { type: string, const: "alma-cert/v1" }
        suite: { type: string, examples: ["ed25519"] }
        issuer: { type: string, description: name@region }
        subject: { type: string, description: name@region }
        schemaId: { type: string }
        claims:
          type: object
          additionalProperties: true
        issuedAt: { type: string, description: "ISO 8601 (deterministic, derived from tick)" }
        signature: { type: string, description: base64 }